+++ analysis/src/Commands.ml
@@ -28,6 +28,10 @@ let completion ~debug ~path ~pos ~currentFile =
     |> List.map Protocol.stringifyCompletionItem
     |> Protocol.array)
 
+let inlayhint ~path ~pos ~maxLength ~debug =
+  let result = Hint.inlay ~path ~pos ~maxLength ~debug |> Protocol.array in
+  print_endline result
+
 let hover ~path ~pos ~currentFile ~debug =
   let result =
     match Cmt.fullFromPath ~path with
@@ -382,6 +386,11 @@ let test ~path =
                                    (Protocol.stringifyRange range)
                                    indent indent newText)))
           | "dia" -> diagnosticSyntax ~path
+          | "hin" -> (
+            let line_start = 0 in
+            let line_end = 6 in
+            print_endline ("Inlay Hint " ^ path ^ " " ^ string_of_int line_start ^ ":" ^ string_of_int line_end);
+            inlayhint ~path ~pos:(line_start, line_end) ~maxLength:"25" ~debug:false)
           | _ -> ());
           print_newline ())
     in