+++ analysis/src/Protocol.ml
@@ -1,6 +1,13 @@
 type position = {line: int; character: int}
 type range = {start: position; end_: position}
 type markupContent = {kind: string; value: string}
+type inlayHint = {
+  position: position;
+  label: string;
+  kind: int;
+  paddingLeft: bool;
+  paddingRight: bool;
+}
 
 type completionItem = {
   label: string;
@@ -127,6 +134,18 @@ let stringifyCodeAction ca =
   Printf.sprintf {|{"title": "%s", "kind": "%s", "edit": %s}|} ca.title
     (codeActionKindToString ca.codeActionKind)
     (ca.edit |> stringifyCodeActionEdit)
+
+let stringifyHint hint =
+  Printf.sprintf
+    {|{
+    "position": %s,
+    "label": "%s",
+    "kind": %i,
+    "paddingLeft": %b,
+    "paddingRight": %b
+}|}
+    (stringifyPosition hint.position)
+    (Json.escape hint.label) hint.kind hint.paddingLeft hint.paddingRight
 
 (* https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#diagnostic *)
 let stringifyDiagnostic d =