+++ client/src/extension.ts
@@ -235,6 +235,17 @@ export function activate(context: ExtensionContext) {
   // Start the client. This will also launch the server
   client.start();
 
+  // Restart the language client automatically when certain configuration
+  // changes. These are typically settings that affect the capabilities of the
+  // language client, and because of that requires a full restart.
+  context.subscriptions.push(
+    workspace.onDidChangeConfiguration(({ affectsConfiguration }) => {
+      if (affectsConfiguration("rescript.settings.inlayHints")) {
+        commands.executeCommand("rescript-vscode.restart_language_server");
+      }
+    })
+  );
+
   // Autostart code analysis if wanted
   if (
     workspace