+++ CHANGELOG.md
@@ -12,6 +12,9 @@
 
 ## master
 
+#### :rocket: New Feature
+
+- Inlay Hints (experimetal). `rescript.settings.inlayHints.enable: true`
 ## v1.4.2
 
 #### :bug: Bug Fix